refactor: extract result submission logic to method
